+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Populate resource correctly in storage rules evaluation. (#4329)
 - Fix bug where invalid CPU was set for 16GiB functions. (#4823)
 - Fix bug where failed function discovery crashed the entire emulator. (#4826)
+- Fix LIST security rule evaluation in storage emulator. (#4827)